+++ databases/alembic/versions/313a6fb99082_update_user_for_command_function_actions.py
@@ -1,0 +1,28 @@
+"""update user for Command Function Actions
+
+Revision ID: 313a6fb99082
+Revises: 03331fc158bc
+Create Date: 2020-12-19 18:24:58.481151
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '313a6fb99082'
+down_revision = '03331fc158bc'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    import sys
+    import os
+    sys.path.append(os.path.abspath(os.path.join(__file__, "../../../..")))
+    from databases.alembic_post_utils import write_revision_post_alembic
+    write_revision_post_alembic(revision)
+
+
+def downgrade():
+    pass