+++ mycodo/outputs/pwm_shell.py
@@ -63,13 +63,8 @@ OUTPUT_INFORMATION = {
         },
         {
             'id': 'linux_command_user',
-            'type': 'select',
-            'default_value': 'pi',
-            'options_select': [
-                ('pi', 'pi'),
-                ('mycodo', 'mycodo'),
-                ('root', 'root')
-            ],
+            'type': 'text',
+            'default_value': 'mycodo',
             'name': lazy_gettext('User'),
             'phrase': lazy_gettext('The user to execute the command')
         },